Move setting menu to pull request list fragment
