+++ app/src/main/java/com/genymobile/pr/ui/MainActivity.java
@@ -1,13 +1,10 @@
 package com.genymobile.pr.ui;
 
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
 
 import com.genymobile.pr.R;
 
@@ -51,23 +48,5 @@ public class MainActivity extends AppCompatActivity implements LoginFragment.Cal
         getSupportFragmentManager().beginTransaction()
                 .replace(R.id.container, fragment)
                 .commit();
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_pr_list, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.getItemId() == R.id.action_settings) {
-            openSettings();
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void openSettings() {
-        startActivity(new Intent(this, SettingsActivity.class));
     }
 }
